requirements.txt
Pillow
face_recognition
requests
selenium
